+++ packages/sdk/CHANGELOG.md
@@ -9,6 +9,8 @@
 - `VerifiablePresentationV1.createAccountClaims` and `VerifiablePresentationV1.createIdentityClaims` to build the
   subject claims used as input for the `VerifiablePresentationV1.create...` functions.
 - `VerifiableCredentialV1` module and corresponding types and associated functionality.
+- `VerificationAuditRecord.createChecked` which creates a verification audit record while verifying the presentation in
+  the context of the verification request.
 
 ### Changed